workers/requirements.txt
httpx==0.25.2
Pillow==10.1.0
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
beautifulsoup4==4.12.2
loguru==0.7.2
requests==2.31.0
